+++ composio/local_tools/file/actions/write_file.py
@@ -1,7 +1,7 @@
 import typing as t
 from pathlib import Path
 from pydantic import BaseModel, Field
-from composio.tools.local import Action
+from composio.core.local import Action
 
 class WriteFileRequest(BaseModel):
     """Safe file request schema."""
