+++ composio/local_tools/local_workspace/cmd_manager/__init__.py
@@ -1,1 +1,1 @@
-from .cmd_manager_tool import CmdManagerTool
+from .tool import CmdManagerTool
