+++ composio/local_tools/local_workspace/history_keeper/__init__.py
@@ -1,1 +1,1 @@
-from .history_keeper_tool import HistoryKeeper
+from .tool import HistoryKeeper
