--- composio/local_tools/local_workspace/history_keeper/history_keeper_tool.py
@@ -1,38 +1,0 @@
-import typing as t
-from typing import Optional
-from composio.tools.local import Tool, Action
-
-from composio.local_tools.local_workspace.commons.get_logger import get_logger
-from composio.local_tools.local_workspace.commons import (
-    WorkspaceManagerFactory,
-    HistoryProcessor
-)
-from .actions import GetWorkspaceHistory
-
-logger = get_logger()
-
-class HistoryKeeper(Tool):
-    """
-    local workspace tool which can maintain history across commands.
-    """
-
-    workspace_factory: Optional[WorkspaceManagerFactory] = None
-    history_processor: Optional[HistoryProcessor] = None
-
-    def actions(self) -> list[t.Type[Action]]:
-        return [GetWorkspaceHistory]
-
-    def triggers(self) -> list:
-        return []
-
-    def set_workspace_factory(self, workspace_factory: WorkspaceManagerFactory):
-        self.workspace_factory = workspace_factory
-
-    def get_workspace_factory(self) -> Optional[WorkspaceManagerFactory]:
-        return self.workspace_factory
-
-    def set_history_processor(self, history_processor: HistoryProcessor):
-        self.history_processor = history_processor
-
-    def get_history_processor(self) -> Optional[HistoryProcessor]:
-        return self.history_processor
