+++ composio/local_tools/local_workspace/workspace/__init__.py
@@ -1,1 +1,1 @@
-from .workspace_tool import LocalWorkspace
+from .tool import LocalWorkspace
