+++ composio/local_tools/local_workspace/workspace/actions/base_workspace_action.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import Optional
 from pydantic import BaseModel
-from composio.tools.local import Action
+from composio.core.local import Action
 
 from composio.local_tools.local_workspace.commons import (
     get_logger,
